A-BasicCommom/powerlawapsp/requirements.txt
cupy-cuda101==8.1.0
numpy==1.18.2
scipy==1.5.4